fix: correct icon, helper constants, consolidated style
