+++ src/assets/DimensionItemIcons/GenericIcon.js
@@ -1,0 +1,15 @@
+import React from 'react'
+
+export default (
+    <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="16"
+        height="16"
+        viewBox="0 0 16 16"
+    >
+        <path
+            fill="#6E7A8A"
+            d="M11,5 L11,11 L5,11 L5,5 L11,5 Z M10,6 L6,6 L6,10 L10,10 L10,6 Z"
+        />
+    </svg>
+)
