--- src/components/DynamicDimension/styles/ItemSelector.style.js
@@ -1,12 +1,0 @@
-import css from 'styled-jsx/css'
-import { spacers, colors } from '@dhis2/ui-core'
-
-export default css`
-    .emptySelection {
-        text-align: center;
-        font-size: 14px;
-        line-height: 16px;
-        margin: ${spacers.dp24} 0 0;
-        color: ${colors.grey700};
-    }
-`
