+++ src/modules/dimensionSelectorHelper.js
@@ -1,0 +1,5 @@
+export const TRANSFER_OPTIONS_WIDTH = '420px'
+
+export const TRANSFER_SELECTED_WIDTH = '298px'
+
+export const TRANSFER_HEIGHT = '512px'
